Move debug functions to debug package
