+++ cmd/extract.go
@@ -29,6 +29,7 @@ import (
 	"strings"
 
 	"github.com/bitcanon/mactool/cli"
+	"github.com/bitcanon/mactool/debug"
 	"github.com/bitcanon/mactool/mac"
 	"github.com/bitcanon/mactool/utils"
 	"github.com/spf13/cobra"
@@ -131,7 +132,7 @@ var extractCmd = &cobra.Command{
 
 		// Print the configuration debug if the --debug flag is set
 		if viper.GetBool("debug") {
-			utils.PrintConfigDebug()
+			debug.PrintConfigDebug()
 		}
 
 		// Extract MAC addresses from string and
